Fix minimum/maximum==0 for UpDownWidget and RangeWidget (#344)
