+++ src/components/widgets/UpDownWidget.js
@@ -1,21 +1,7 @@
 import React, { PropTypes } from "react";
 
+import { rangeSpec } from "../../utils";
 import BaseInput from "./BaseInput";
-
-
-function rangeSpec(schema) {
-  const spec = {};
-  if (schema.multipleOf) {
-    spec.step = schema.multipleOf;
-  }
-  if (schema.minimum) {
-    spec.min = schema.minimum;
-  }
-  if (schema.maximum) {
-    spec.max = schema.maximum;
-  }
-  return spec;
-}
 
 function UpDownWidget(props) {
   return <BaseInput type="number" {...props} {...rangeSpec(props.schema)} />;
